+++ SimpePrimitiveWizards_73_44-QA/0x006f/BhavOperandWiz0x006f.UI.cs
@@ -113,7 +113,7 @@ namespace whse.PrimitiveWizards.Wiz0x006f
 
                 wrapper.ProcessData(entryArray[0].PFD, entryArray[0].Package);
 
-                ushort strIndex = (ushort)new pjse.StrChooser(true).Strnum(wrapper);
+                int strIndex = new pjse.StrChooser(true).Strnum(wrapper);
 
                 if (strIndex >= 0)
                 {
