+++ SimpePrimitiveWizards_73_44-QA/0x0072/BhavOperandWiz0x0072.UI.cs
@@ -133,14 +133,14 @@ namespace whse.PrimitiveWizards.Wiz0x0072
 
                 wrapper.ProcessData(entryArray[0].PFD, entryArray[0].Package);
 
-                ushort strIndex = (ushort)new pjse.StrChooser(true).Strnum(wrapper);
+                int strIndex = new pjse.StrChooser(true).Strnum(wrapper);
 
                 if (strIndex >= 0)
                 {
                     bool internalchg = this.internalchg;
                     this.internalchg = true;
 
-                    WizardHelpers.SetValue(textLocoAnimEntry, strIndex, checkDecimal);
+                    WizardHelpers.SetValue(textLocoAnimEntry, (ushort)strIndex, checkDecimal);
                     UpdateLocoAnimName();
 
                     this.internalchg = internalchg;
