+++ SimpePrimitiveWizards_73_44-QA/0x007e/BhavOperandWiz0x007e.UI.cs
@@ -125,7 +125,7 @@ namespace whse.PrimitiveWizards.Wiz0x007e
 
                 wrapper.ProcessData(entryArray[0].PFD, entryArray[0].Package);
 
-                ushort strIndex = (ushort)new pjse.StrChooser(true).Strnum(wrapper);
+                int strIndex = new pjse.StrChooser(true).Strnum(wrapper);
 
                 if (strIndex >= 0)
                 {
